Add Custom Icons to Icon Generator (#65)
The desire here is to be able to reference custom icons in addition to material icons with the #icon helper method. Currently, material icons can be rendered with = icon(<name>). The goal is to have the ability to similarly render custom icons that live inside app/javascript/images/icons

This ports over two webpack loaders from sayfa to optimize svg files and allow the ability to override the color. One of the goals is to have the icons behave similar to material icons in that, if they don't have a specified color, they will inherit the font color of the parent css element.